Instantiate keyMap object by function.  Change menu state to being held in object
Create a function to set the keyMap object.  This allows control over
when the object is set so that other functionality/functions (namely the
menus) can be instantiated first.
Menu state is now held in an array so that submenus can just push their
name and previous menu actions can just pop off the last name.  A
function getCurrMenuObj uses the array to find the correct menu object
(from the screenMenus object) for screen dispaly.