Create stack operation function to reduce repetitive code
Operator keys, and others, were having a specific function call per key
to simply execute the operation with the correct arguments.  Create
stackOperation function that takes a the number of lines on which to
operate and the callback function.  This function takes care of getting
the arguments to the callback and displaying the results.